data-generator/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3 